Allow for a configurable PAN ID
Up to now, the code would chose the lower 16 bits of the IEEE
address of the Zigbee dongle as the desired PAN ID.
This commit changes the default to the "I don't care" value
(0xFFFF) for the ZStack. This way, the dongle will
- use the established one if there is one, or
- create a random new one.

Also, a new configuration argument is added to specify a
desired PAN ID explicitly.