add support for ZDO_MGMT_DIRECT_JOIN_REQ call
This allows to create join information in the coordinator.
Send any message to MQTT topic <prefix>/write/directjoin/<IEEEAddr>
where the <IEEEAddr> is the, well, IEEE address (64bit) of the
device you want to join.

Comes in handy when you have reset the NVRAM of your coordinator
(i.e. the USB stick) and you have some exceptionally hard to join
devices that you want to push back into the network.